# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certiwise",
        "description": "Api para generar certificados",
        "contact": {
            "email": "contact@academlo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/{tenant}/collections": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "Obtener lista de colecciones",
                "operationId": "a85ba11d8ba1174932b9aa1507d20716",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de colecciones por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de colecciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de colecciones"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Collection"
                ],
                "summary": "Crear una nueva colección",
                "operationId": "bec0512c5aff5cdce3188bc90086b110",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Colección de muestra"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción de la colección"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Colección creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Colección creada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/v1/{tenant}/collections/{id}": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "Obtener detalles de una colección",
                "operationId": "0122501633da4d66a11aff41e5610f88",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colección",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la colección",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Detalles de la colección"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Colección no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Collection"
                ],
                "summary": "Actualizar una colección existente",
                "operationId": "a385550b22c742e854413ff00f0624d5",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colección",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Colección actualizada"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada de la colección"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Colección actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Colección actualizada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Colección no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Collection"
                ],
                "summary": "Eliminar una colección",
                "operationId": "a6ae07ada5a278d8450c1f543d6ea985",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colección",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Colección eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Colección eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Colección no encontrada"
                    }
                }
            }
        },
        "/api/v1/{tenant}/collections/{id}/events": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "Obtener eventos de una colección",
                "operationId": "7eb7fcce3201ed3b5d1b1486b37135f0",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colección",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de eventos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de eventos de la colección paginados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de eventos de la colección paginados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Colección no encontrada"
                    }
                }
            }
        },
        "/api/v1/{tenant}/collections/{id}/add-event": {
            "post": {
                "tags": [
                    "Collection"
                ],
                "summary": "Agregar un evento a una colección",
                "operationId": "c651a84a5ed520154894e544ca85ae0b",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Academlo"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colección",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento agregado exitosamente a la colección",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evento agregado exitosamente a la colección"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Colección o evento no encontrado"
                    }
                }
            }
        },
...
    }
}